guardian_monitor/requirements.txt
langgraph
langchain
langchain-openai
paramiko
pydantic
python-dotenv
langchain-ollama
python-telegram-bot
duckduckgo-search
